style: auth를 기능(dashboard, default, memo)로 분리하여 코드 저장
