+++ src/utils/auth/dashboard/checkValid.ts
@@ -1,0 +1,14 @@
+import { errorResponse, isValidResponse } from 'utils/auth';
+import { getMemosResponse, getTagsResponse } from './interface';
+
+export const isGetMemosResponse = (
+  response: getMemosResponse | errorResponse
+): response is getMemosResponse => {
+  return isValidResponse(response as getMemosResponse);
+};
+
+export const isGetTagsResponse = (
+  response: getTagsResponse | errorResponse
+): response is getTagsResponse => {
+  return (response as getTagsResponse).tags !== null;
+};
