+++ src/utils/auth/memo/checkValid.ts
@@ -1,0 +1,26 @@
+import { errorResponse, validResponse, isValidResponse } from 'utils/auth';
+import { cuMemoResponse, searchMemoResponse } from './interface';
+
+export const isSearchMemoResponse = (
+  response: searchMemoResponse | errorResponse
+): response is searchMemoResponse => {
+  return (response as searchMemoResponse).text !== null;
+};
+
+export const isCreateMemoResponse = (
+  response: cuMemoResponse | errorResponse
+): response is cuMemoResponse => {
+  return (response as cuMemoResponse).content !== null;
+};
+
+export const isUpdateMemoResponse = (
+  response: cuMemoResponse | errorResponse
+): response is cuMemoResponse => {
+  return (response as cuMemoResponse).content !== null;
+};
+
+export const isDeleteMemoResponse = (
+  response: validResponse | errorResponse
+): response is validResponse => {
+  return isValidResponse(response);
+};
